app.py
from dotenv import load_dotenv
import os

from flask import Flask, request, jsonify
import mysql.connector

# Load environment variables from .env file
load_dotenv()

app = Flask(__name__)

db_host = os.getenv('HOST')
db_username = os.getenv('USERNAME')
db_password = os.getenv('PASSWORD')
db = 'COMP3161_Group_Project'


@app.route('/register', methods=['POST'])
def register_user():
    try:
        # Implement registration logic here
        return jsonify(message="User registered successfully"), 201
    except ValueError as e:
        return jsonify(error=str(e)), 400
    except Exception as e:
        return jsonify(error="An unexpected error occurred"), 500

@app.route('/login', methods=['POST'])
def login():
    try:
        # Implement login logic here
        return jsonify(message="Login successful"), 200
    except Exception as e:
        return jsonify(error="Login failed"), 401

@app.route('/course', methods=['POST'])
def create_course():
    try:
        # Implement course creation logic here
        return jsonify(message="Course created successfully"), 201
    except Exception as e:
        return jsonify(error="Failed to create course"), 500

@app.route('/courses', methods=['GET'])
def get_courses():
    try:
        # Implement logic to retrieve all courses
        return jsonify(courses=[]), 200
    except Exception as e:
        return jsonify(error="Failed to retrieve courses"), 500

@app.route('/student/courses', methods=['GET'])
def get_student_courses():
    try:
        # Implement logic to retrieve courses for a particular student
        return jsonify(courses=[]), 200
    except Exception as e:
        return jsonify(error="Failed to retrieve student's courses"), 500

@app.route('/lecturer/courses', methods=['GET'])
def get_lecturer_courses():
    try:
        # Implement logic to retrieve courses taught by a particular lecturer
        return jsonify(courses=[]), 200
    except Exception as e:
        return jsonify(error="Failed to retrieve lecturer's courses"), 500

@app.route('/course/register', methods=['POST'])
def register_for_course():
    try:
        # Implement course registration logic here
        return jsonify(message="Registered for course successfully"), 200
    except Exception as e:
        return jsonify(error="Failed to register for course"), 500

@app.route('/course/members', methods=['GET'])
def get_course_members():
    try:
        # Implement logic to retrieve members of a particular course
        return jsonify(members=[]), 200
    except Exception as e:
        return jsonify(error="Failed to retrieve course members"), 500

@app.route('/course/calendar', methods=['GET', 'POST'])
def calendar_events():
    try:
        if request.method == 'POST':
            # Implement logic to create a calendar event for a course
            return jsonify(message="Calendar event created successfully"), 201
        # Implement logic to retrieve all calendar events for a course
        return jsonify(events=[]), 200
    except Exception as e:
        return jsonify(error="Failed to process calendar events"), 500

@app.route('/student/calendar', methods=['GET'])
def student_calendar_events():
    try:
        # Implement logic to retrieve all calendar events for a student on a particular date
        return jsonify(events=[]), 200
    except Exception as e:
        return jsonify(error="Failed to retrieve student's calendar events"), 500

@app.route('/course/forum', methods=['GET', 'POST'])
def forums():
    try:
        if request.method == 'POST':
            # Implement logic to create a forum for a particular course
            return jsonify(message="Forum created successfully"), 201
        # Implement logic to retrieve all forums for a particular course
        return jsonify(forums=[]), 200
    except Exception as e:
        return jsonify(error="Failed to process forum requests"), 500

@app.route('/forum/thread', methods=['GET', 'POST'])
def discussion_thread():
    try:
        if request.method == 'POST':
            # Implement logic to add a new discussion thread to a forum
            return jsonify(message="Discussion thread added successfully"), 201
        # Implement logic to retrieve all discussion threads for a particular forum
        return jsonify(threads=[]), 200
    except Exception as e:
        return jsonify(error="Failed to process discussion thread requests"), 500

@app.route('/thread/reply', methods=['POST'])
def reply_to_thread():
    try:
        # Implement logic to reply to a thread (replies can have replies)
        return jsonify(message="Reply added successfully"), 201
    except Exception as e:
        return jsonify(error="Failed to add reply to thread"), 500

@app.route('/course/content', methods=['GET', 'POST'])
def course_content():
    try:
        if request.method == 'POST':
            # Implement logic to add course content
            return jsonify(message="Content added successfully"), 201
        # Retrieve all course content for a particular course
        return jsonify(content=[]), 200
    except Exception as e:
        return jsonify(error="Failed to process course content requests"), 500

@app.route('/assignment/submit', methods=['POST'])
def submit_assignment():
    try:
        # Implement logic for a student submitting an assignment
        return jsonify(message="Assignment submitted successfully"), 201
    except Exception as e:
        return jsonify(error="Failed to submit assignment"), 500

@app.route('/assignment/grade', methods=['POST'])
def grade_assignment():
    try:
        # Implement logic for a lecturer grading an assignment
        return jsonify(message="Grade submitted successfully"), 201
    except Exception as e:
        return jsonify(error="Failed to submit grade"), 500


@app.errorhandler(404)
def resource_not_found(e):
    return jsonify(error="Resource not found"), 404

@app.errorhandler(500)
def internal_server_error(e):
    return jsonify(error="Internal server error"), 500
    

if __name__ == '__main__':
    app.run(port=8080, debug=True)